Major breaking changes to Session
Session had issues with object data types -- it could only work with STRING values for the Value object. NOW it works with all types doing proper serilization as necessary

This means it breaks session["key"]?.Value -- if you are expecting VALUE to be a string

Implemented the ToString() on session["key"] AND also its Value to provide string results if you expect them. Otherwise it returns the object desired

Auto Commit was only happning if app.UseSession(0 was available in program.cs -- but this new implementation no longer works with app.UseSession and therefore auto-commit went away. NOW implemented that functionality too.

Still maintains cookies or local data to remember SessionID across session instances so long as it has not expired on the local machine or in the table store.