OBrown solved a BIG BUG Issue with Getting Data from the Tables.
Read and Write  Entity had an issue with Try Catch and also converting data correctly. The rules somehow changed when I moved away from the Activator Model to the pagable model for table query execution. It removed the ReadEntity flow specifically for the GetCollectionCore function, which is the primary data collection for Data Access. That pageable model actually somehow changes the nature of the data so that even the Convert function that ensure dateTime and boolean values and others are formatted correctly -- FAILED. Had to rewrite the entire chain of those functions to double check values to make sure they are all converted correctly.

Still don't understand how SESSION was able to function correctly and NOT any other data grab? Setting data was fine, but it was bringing it back incorrectly. Now it works. Noticed it first with duplicate data in the table for no reason. The looked at values. Then was really tipped off when a bool was always coming back as false but was clearly true in the db.