OBrown spotted a bug in Lambda expressions that was a showstopper. Simple expressions were returning false positives that led to duplicate data or wrong data and in some cases data that didnt belong to the query.
Resolved issues with the expression builders for both Blob and Table data. Moved the Session Object to the SessionManagement.cs to keep it out of harms way in all of the changes.  Will continue testing and improving here because complex lambda's still fail defeating the purpose and intention.

Saving working and improved code